+++ executor/executor.go
@@ -57,7 +57,6 @@ var (
 	_ Executor = &TableScanExec{}
 	_ Executor = &TopNExec{}
 	_ Executor = &UnionExec{}
-	_ Executor = &RecoverIndexExec{}
 	_ Executor = &CheckIndexExec{}
 	_ Executor = &HashJoinExec{}
 	_ Executor = &IndexLookUpExecutor{}
