+++ executor/executor_test.go
@@ -213,179 +213,6 @@ func (s *testSuite) TestAdmin(c *C) {
 	c.Assert(err, NotNil)
 }
 
-func findIndexByName(idxName string, indices []*model.IndexInfo) *model.IndexInfo {
-	for _, idx := range indices {
-		if idx.Name.L == idxName {
-			return idx
-		}
-	}
-	return nil
-}
-
-func (s *testSuite) TestAdminRecoverIndex(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists admin_test")
-	tk.MustExec("create table admin_test (c1 int, c2 int, c3 int default 1, index (c1), unique key(c2))")
-	tk.MustExec("insert admin_test (c1, c2) values (1, 1), (2, 2), (NULL, NULL)")
-
-	r := tk.MustQuery("admin recover index admin_test c1")
-	r.Check(testkit.Rows("0 3"))
-
-	r = tk.MustQuery("admin recover index admin_test c2")
-	r.Check(testkit.Rows("0 3"))
-
-	tk.MustExec("admin check index admin_test c1")
-	tk.MustExec("admin check index admin_test c2")
-
-	tk.MustExec("drop table if exists admin_test")
-	tk.MustExec("create table admin_test (c1 int, c2 int, c3 int default 1, primary key(c1), unique key(c2))")
-	tk.MustExec("insert admin_test (c1, c2) values (1, 1), (2, 2), (3, 3), (10, 10), (20, 20)")
-	// pk is handle, no additional unique index, no way to recover
-	_, err := tk.Exec("admin recover index admin_test c1")
-	// err:index is not found
-	c.Assert(err, NotNil)
-
-	r = tk.MustQuery("admin recover index admin_test c2")
-	r.Check(testkit.Rows("0 5"))
-	tk.MustExec("admin check index admin_test c2")
-
-	// Make some corrupted index.
-	s.ctx = mock.NewContext()
-	s.ctx.Store = s.store
-	is := s.domain.InfoSchema()
-	dbName := model.NewCIStr("test")
-	tblName := model.NewCIStr("admin_test")
-	tbl, err := is.TableByName(dbName, tblName)
-	c.Assert(err, IsNil)
-
-	tblInfo := tbl.Meta()
-	idxInfo := findIndexByName("c2", tblInfo.Indices)
-	indexOpr := tables.NewIndexWithBuffer(tblInfo, idxInfo)
-	sc := s.ctx.GetSessionVars().StmtCtx
-	txn, err := s.store.Begin()
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums(1), 1)
-	c.Assert(err, IsNil)
-	err = txn.Commit(context.Background())
-	c.Assert(err, IsNil)
-	_, err = tk.Exec("admin check table admin_test")
-	c.Assert(err, NotNil)
-	_, err = tk.Exec("admin check index admin_test c2")
-	c.Assert(err, NotNil)
-
-	r = tk.MustQuery("SELECT COUNT(*) FROM admin_test USE INDEX(c2)")
-	r.Check(testkit.Rows("4"))
-
-	r = tk.MustQuery("admin recover index admin_test c2")
-	r.Check(testkit.Rows("1 5"))
-
-	r = tk.MustQuery("SELECT COUNT(*) FROM admin_test USE INDEX(c2)")
-	r.Check(testkit.Rows("5"))
-	tk.MustExec("admin check index admin_test c2")
-	tk.MustExec("admin check table admin_test")
-
-	txn, err = s.store.Begin()
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums(10), 10)
-	c.Assert(err, IsNil)
-	err = txn.Commit(context.Background())
-	c.Assert(err, IsNil)
-
-	_, err = tk.Exec("admin check index admin_test c2")
-	c.Assert(err, NotNil)
-	r = tk.MustQuery("admin recover index admin_test c2")
-	r.Check(testkit.Rows("1 5"))
-	tk.MustExec("admin check index admin_test c2")
-	tk.MustExec("admin check table admin_test")
-
-	txn, err = s.store.Begin()
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums(1), 1)
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums(2), 2)
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums(3), 3)
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums(10), 10)
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums(20), 20)
-	c.Assert(err, IsNil)
-	err = txn.Commit(context.Background())
-	c.Assert(err, IsNil)
-
-	_, err = tk.Exec("admin check table admin_test")
-	c.Assert(err, NotNil)
-	_, err = tk.Exec("admin check index admin_test c2")
-	c.Assert(err, NotNil)
-
-	r = tk.MustQuery("SELECT COUNT(*) FROM admin_test USE INDEX(c2)")
-	r.Check(testkit.Rows("0"))
-
-	r = tk.MustQuery("SELECT COUNT(*) FROM admin_test")
-	r.Check(testkit.Rows("5"))
-
-	r = tk.MustQuery("admin recover index admin_test c2")
-	r.Check(testkit.Rows("5 5"))
-
-	r = tk.MustQuery("SELECT COUNT(*) FROM admin_test USE INDEX(c2)")
-	r.Check(testkit.Rows("5"))
-
-	tk.MustExec("admin check index admin_test c2")
-	tk.MustExec("admin check table admin_test")
-}
-
-func (s *testSuite) TestAdminRecoverIndex1(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.ctx = mock.NewContext()
-	s.ctx.Store = s.store
-	dbName := model.NewCIStr("test")
-	tblName := model.NewCIStr("admin_test")
-	sc := s.ctx.GetSessionVars().StmtCtx
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists admin_test")
-	tk.MustExec("create table admin_test (c1 varchar(255), c2 int, c3 int default 1, primary key(c1), unique key(c2))")
-	tk.MustExec("insert admin_test (c1, c2) values ('1', 1), ('2', 2), ('3', 3), ('10', 10), ('20', 20)")
-
-	r := tk.MustQuery("SELECT COUNT(*) FROM admin_test USE INDEX(`primary`)")
-	r.Check(testkit.Rows("5"))
-
-	is := s.domain.InfoSchema()
-	tbl, err := is.TableByName(dbName, tblName)
-	c.Assert(err, IsNil)
-
-	tblInfo := tbl.Meta()
-	idxInfo := findIndexByName("primary", tblInfo.Indices)
-	c.Assert(idxInfo, NotNil)
-	indexOpr := tables.NewIndexWithBuffer(tblInfo, idxInfo)
-
-	txn, err := s.store.Begin()
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums("1"), 1)
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums("2"), 2)
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums("3"), 3)
-	c.Assert(err, IsNil)
-	err = indexOpr.Delete(sc, txn, types.MakeDatums("10"), 4)
-	c.Assert(err, IsNil)
-	err = txn.Commit(context.Background())
-	c.Assert(err, IsNil)
-
-	r = tk.MustQuery("SELECT COUNT(*) FROM admin_test USE INDEX(`primary`)")
-	r.Check(testkit.Rows("1"))
-
-	r = tk.MustQuery("admin recover index admin_test `primary`")
-	r.Check(testkit.Rows("4 5"))
-
-	r = tk.MustQuery("SELECT COUNT(*) FROM admin_test USE INDEX(`primary`)")
-	r.Check(testkit.Rows("5"))
-
-	tk.MustExec("admin check table admin_test")
-	tk.MustExec("admin check index admin_test c2")
-	tk.MustExec("admin check index admin_test `primary`")
-}
-
 func (s *testSuite) fillData(tk *testkit.TestKit, table string) {
 	tk.MustExec("use test")
 	tk.MustExec(fmt.Sprintf("create table %s(id int not null default 1, name varchar(255), PRIMARY KEY(id));", table))
